F01: Add implementation of VillageService
