+++ src/main/java/com/carto/sn/service/VillageService.java
@@ -1,5 +1,10 @@
 package com.carto.sn.service;
 
-public interface VillageService {
+import java.util.List;
 
+import com.carto.sn.entities.Village;
+
+public interface VillageService {
+	public List <Village> tousLesVillages();
+	public List <Village> findVillageByNomCommune(String nomCommune);
 }
